# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: NexusBridge HTTP API
  version: 0.1.0
  description: |
    NexusBridge 后端 HTTP API 契约。当前文件描述已实现接口，前端新需求先写入 docs/api/requests.md。
servers:
  - url: http://0.0.0.0:8090
    description: 本地默认服务
...
paths:
...
  /api/logs:
    get:
      tags: [logs]
      summary: 获取当前进程最近日志
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 2000
            default: 500
      responses:
        "200":
          description: 按时间从旧到新排列的日志快照
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogSnapshot"
        "400":
          $ref: "#/components/responses/Error"
  /api/logs/stream:
    get:
      tags: [logs]
      summary: 实时订阅当前进程日志
      description: 先发送 snapshot 事件，再发送后续 log 事件；服务端定期发送 SSE 注释保持连接。
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 2000
            default: 500
      responses:
        "200":
          description: Server-Sent Events 日志流
          content:
            text/event-stream:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/Error"
...
  /api/session/login:
    post:
      tags: [session]
      summary: 本地登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /api/sites:
    get:
      tags: [sites]
      summary: 获取站点列表
      responses:
        "200":
          description: 站点列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Site"
        "500":
          $ref: "#/components/responses/Error"
  /api/sites/{site_id}/credential:
    parameters:
      - $ref: "#/components/parameters/SiteID"
    get:
      tags: [sites]
      summary: 获取站点凭据状态
      responses:
        "200":
          description: 站点凭据状态，不返回 cookie 明文
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteCredential"
        "404":
          $ref: "#/components/responses/Error"
    post:
      tags: [sites]
      summary: 保存站点 cookie 和 user-agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteCredential"
      responses:
        "200":
          description: 已保存的凭据状态，不返回 cookie 明文
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteCredential"
        "400":
          $ref: "#/components/responses/Error"
  /api/sites/{site_id}/attendance:
    parameters:
      - $ref: "#/components/parameters/SiteID"
    get:
      tags: [sites]
      summary: 获取站点自动签到配置
      responses:
        "200":
          description: 已保存配置，或关闭且未配置时区的默认配置
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteAttendance"
        "404":
          $ref: "#/components/responses/Error"
    post:
      tags: [sites]
      summary: 保存站点自动签到配置
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteAttendanceInput"
      responses:
        "200":
          description: 已保存配置
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteAttendance"
        "400":
          $ref: "#/components/responses/Error"
  /api/sites/{site_id}/fetch:
    parameters:
      - $ref: "#/components/parameters/SiteID"
    post:
      tags: [sites]
      summary: 创建统一的站点分页扫描任务
      parameters:
        - name: trigger
          in: query
          required: false
          description: 媒体页自动触发时传 homepage，其余请求可省略。
          schema:
            type: string
            enum: [manual, homepage]
            default: manual
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteFetchRequest"
      responses:
        "202":
          description: 已创建任务，或返回同站点现有活动任务
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteFetchJob"
        "400":
          $ref: "#/components/responses/Error"
  /api/site-fetch-jobs:
    get:
      tags: [sites, subscriptions]
      summary: 获取持久化站点扫描任务
      parameters:
        - name: site_id
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        "200":
          description: 最近扫描任务
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SiteFetchJob"
        "500":
          $ref: "#/components/responses/Error"
  /api/sites/{site_id}/schedule:
    parameters:
      - $ref: "#/components/parameters/SiteID"
    get:
      tags: [sites, subscriptions]
      summary: 获取站点周期抓取配置
      responses:
        "200":
          description: 已保存配置，或关闭且间隔为 15 分钟的默认配置
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSchedule"
        "400":
          $ref: "#/components/responses/Error"
    post:
      tags: [sites, subscriptions]
      summary: 保存站点周期抓取配置
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteScheduleInput"
      responses:
        "200":
          description: 已保存配置
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSchedule"
        "400":
          $ref: "#/components/responses/Error"
  /api/sites/{site_id}/filter-options:
    parameters:
      - $ref: "#/components/parameters/SiteID"
    get:
      tags: [sites, rules]
      summary: 获取单站点规则编辑器的本地筛选选项
      responses:
        "200":
          description: 分类和标签来自已抓取种子，促销来自站点检索定义
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuleFilterOptions"
        "400":
          $ref: "#/components/responses/Error"
  /api/sites/{site_id}/media-filter-options:
    parameters:
      - $ref: "#/components/parameters/SiteID"
    get:
      tags: [sites, torrents]
      summary: 获取单站点媒体筛选定义
      responses:
        "200":
          description: 完整 checkbox 分组和 role=promotion 的 spstate 定义
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaFilterOptions"
        "404":
          $ref: "#/components/responses/Error"
  /api/torrents:
    get:
      tags: [torrents]
      summary: 分页查询本地种子缓存
      parameters:
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: site_id
          in: query
          required: false
          schema:
            type: string
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: category
          in: query
          required: false
          description: 可重复的分类值；仅与单个 site_id 一起使用，同类值按 OR 匹配。
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
        - name: promotion
          in: query
          required: false
          description: 可重复的 spstate filter_value；仅与单个 site_id 一起使用，同类值按 OR 匹配。
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
        - name: site_checkbox
          in: query
          required: false
          description: 可重复的 group:value；仅与单个 site_id 一起使用，同组值按 OR、不同组按 AND 匹配。
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
        - name: qb_task
          in: query
          required: false
          description: 按本地种子是否关联当前 qB 任务筛选；列表请求不会主动访问 qB。
          schema:
            type: string
            enum: [present, absent]
        - name: qb_progress
          in: query
          required: false
          description: 仅与 qb_task=present 一起使用；complete 为 progress >= 1。
          schema:
            type: string
            enum: [complete, incomplete]
        - name: include_pinned
          in: query
          required: false
          schema:
            type: boolean
            default: true
        - name: sort_by
          in: query
          required: false
          schema:
            type: string
            enum: [source_order, published_at, size_bytes, seeders, leechers, snatches]
            default: published_at
        - name: sort_direction
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: 筛选、排序和范围截取后的种子页
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TorrentPage"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          description: progress 筛选所需的 qB 运行态尚未完成首次同步
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/Error"
  /api/playback/torrents:
    get:
      tags: [torrents, qbittorrent]
      summary: 查询含有完整音频或视频文件的其他 qB 种子
      parameters:
        - name: exclude_site_id
          in: query
          required: false
          schema:
            type: string
        - name: exclude_torrent_id
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        "200":
          description: 按站点发布时间倒序的其他可播放种子
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlaybackTorrent"
        "400":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
  /api/torrents/{site_id}/{torrent_id}/playback:
    parameters:
      - $ref: "#/components/parameters/SiteID"
      - $ref: "#/components/parameters/TorrentID"
    get:
      tags: [torrents, qbittorrent]
      summary: 获取当前种子的实时媒体播放清单
      parameters:
        - name: file
          in: query
          required: false
          description: 要选中的 qB 相对文件名；省略时使用默认媒体。
          schema:
            type: string
      responses:
        "200":
          description: 种子元数据、qB 状态、媒体选集和默认文件
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybackContext"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
  /api/playback/qb/{hash}:
    parameters:
      - name: hash
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [qbittorrent]
      summary: 按 qB hash 获取实时媒体播放上下文
      parameters:
        - name: file
          in: query
          required: false
          description: 要选中的 qB 相对文件名；省略时使用默认媒体。
          schema:
            type: string
      responses:
        "200":
          description: qB 选集、当前文件、所在目录及可选数据库种子详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybackContext"
        "404":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
  /api/media-libraries:
    get:
      tags: [media-libraries]
      summary: 获取全部媒体库节点
      responses:
        "200":
          description: 使用 parent_id 组成层级的媒体库摘要；不触发磁盘扫描。
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MediaLibrarySummary"
    post:
      tags: [media-libraries]
      summary: 创建媒体库节点
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaLibrarySaveRequest"
      responses:
        "201":
          description: 创建后的媒体库及首次扫描结果。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaLibraryDetail"
        "400":
          $ref: "#/components/responses/Error"
  /api/media-libraries/{library_id}:
    parameters:
      - name: library_id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [media-libraries]
      summary: 获取媒体库缓存详情
      responses:
        "200":
          description: 媒体库详情；不触发磁盘扫描。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaLibraryDetail"
        "404":
          $ref: "#/components/responses/Error"
    put:
      tags: [media-libraries]
      summary: 更新媒体库层级、目录和本地设置
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaLibrarySaveRequest"
      responses:
        "200":
          description: 更新后的媒体库及扫描结果。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaLibraryDetail"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [media-libraries]
      summary: 删除叶子媒体库配置和缓存
      description: 有子媒体库时返回 409；不删除目录或任何源文件。
      responses:
        "200":
          description: 删除结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/media-libraries/{library_id}/scan:
    parameters:
      - name: library_id
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [media-libraries]
      summary: 手动重扫媒体库
      description: 递归扫描当前节点目录，并跳过直接子媒体库负责的目录树。
      responses:
        "200":
          description: 更新后的媒体库详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaLibraryDetail"
        "404":
          $ref: "#/components/responses/Error"
  /api/series:
    get:
      tags: [series]
      summary: 获取全部剧集缓存摘要
      responses:
        "200":
          description: 剧集目录、视频数量和上次扫描状态；不触发磁盘扫描。
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SeriesSummary"
    post:
      tags: [series]
      summary: 创建并扫描本地剧集
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeriesSaveRequest"
      responses:
        "201":
          description: 创建后的剧集详情和首次扫描结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeriesDetail"
        "400":
          $ref: "#/components/responses/Error"
  /api/series/{series_id}:
    parameters:
      - name: series_id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [series]
      summary: 获取一个剧集的缓存详情
      responses:
        "200":
          description: 剧集详情；不触发磁盘扫描。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeriesDetail"
        "404":
          $ref: "#/components/responses/Error"
    put:
      tags: [series]
      summary: 更新并扫描本地剧集
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeriesSaveRequest"
      responses:
        "200":
          description: 更新后的剧集详情和扫描结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeriesDetail"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [series]
      summary: 删除剧集配置和缓存
      description: 不删除目录或任何源文件。
      responses:
        "200":
          description: 删除结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "404":
          $ref: "#/components/responses/Error"
  /api/series/{series_id}/scan:
    parameters:
      - name: series_id
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [series]
      summary: 手动重扫剧集目录
      responses:
        "200":
          description: 更新后的剧集详情；不可读目录保留旧缓存并标为不可用。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeriesDetail"
        "404":
          $ref: "#/components/responses/Error"
  /api/series/{series_id}/selection:
    parameters:
      - name: series_id
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [series]
      summary: 保存剧集最后选择的视频
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeriesSelectionRequest"
      responses:
        "200":
          description: 更新后的剧集详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeriesDetail"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/series/{series_id}/playback:
    parameters:
      - name: series_id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [series]
      summary: 重扫剧集并获取统一播放上下文
      parameters:
        - name: path
          in: query
          required: false
          description: 优先选择的剧集视频绝对路径。
          schema:
            type: string
      responses:
        "200":
          description: 带剧集摘要和有序视频清单的播放上下文
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybackContext"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
  /api/playback/file:
    get:
      tags: [files, qbittorrent]
      summary: 按本机文件自动识别播放上下文
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 自动归一到数据库种子、qB 任务或单文件播放上下文
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybackContext"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /api/playback/external-subtitle:
    get:
      tags: [files]
      summary: 返回视频关联的外挂字幕产物
      parameters:
        - { name: path, in: query, required: true, schema: { type: string } }
        - { name: format, in: query, required: false, schema: { type: string, enum: [vtt, ass] } }
      responses:
        "200": { description: WebVTT 或 ASS 字幕内容 }
        "404": { $ref: "#/components/responses/Error" }
    put:
      tags: [files]
      summary: 保存或替换视频的外挂字幕关联
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaybackExternalSubtitleRequest"
      responses:
        "200":
          description: 已保存的关联
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlaybackExternalSubtitle" }
        "400": { $ref: "#/components/responses/Error" }
    delete:
      tags: [files]
      summary: 删除视频的外挂字幕关联但保留源文件
      parameters:
        - { name: path, in: query, required: true, schema: { type: string } }
      responses:
        "204": { description: 已删除关联 }
...
  /api/playback/qb/{hash}/media/{file_index}:
    parameters:
      - name: hash
        in: path
        required: true
        schema:
          type: string
      - name: file_index
        in: path
        required: true
        schema:
          type: integer
          minimum: 0
    get:
      tags: [qbittorrent]
      summary: 按 qB hash 和文件索引传输源媒体
      responses:
        "200":
          description: 完整文件或 Range 响应
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "206":
          description: 字节范围响应
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
    head:
      tags: [qbittorrent]
      summary: 查询 qB 源媒体响应头
      responses:
        "200":
          description: 文件长度、类型、修改时间和 Range 能力
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
...
components:
  parameters:
    SiteID:
      name: site_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Error:
      description: 错误响应
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    LogSnapshot:
      type: object
      required: [items, total, capacity]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/LogEntry"
        total:
          type: integer
          minimum: 0
          description: 当前环形缓冲区内的日志数量。
        capacity:
          type: integer
          minimum: 0
          description: 当前进程日志缓冲区容量。
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    LoginResponse:
      type: object
      required: [token]
      properties:
        token:
          type: string
    Site:
      type: object
      required: [id, name, base_url, has_cookie]
      properties:
        id:
          type: string
        name:
          type: string
        base_url:
          type: string
        user_agent:
          type: string
        has_cookie:
          type: boolean
    SiteCredential:
      type: object
      required: [site_id, base_url, user_agent, has_cookie]
      properties:
        site_id:
          type: string
        base_url:
          type: string
        user_agent:
          type: string
        cookie:
          type: string
          description: 只允许请求时提交；响应会清空。
        has_cookie:
          type: boolean
...